+++ src/frontend/src/optimizer/rule/apply_to_join.rs
@@ -12,11 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use crate::optimizer::max_one_row_visitor::MaxOneRowVisitor;
 use super::{BoxedRule, Rule};
-use crate::optimizer::plan_node::{
-    LogicalAgg, LogicalApply, LogicalExpand, LogicalHopWindow, LogicalJoin, LogicalProjectSet,
-    LogicalTopN, PlanTreeNodeUnary,
-};
+use crate::optimizer::plan_node::{LogicalApply, LogicalJoin};
 use crate::optimizer::plan_visitor::PlanVisitor;
 use crate::optimizer::PlanRef;
 
@@ -33,12 +31,9 @@ impl Rule for ApplyToJoinRule {
             return None;
         }
 
-        if max_one_row {
-            // Try to eliminate max one row restoration.
-            let mut max_one_row_visitor = MaxOneRowVisitor();
-            if !max_one_row_visitor.visit(right.clone()) {
-                return None;
-            }
+        // Try to eliminate max one row.
+        if max_one_row && !MaxOneRowVisitor().visit(right.clone()) {
+            return None;
         }
 
         Some(LogicalJoin::new(left, right, join_type, on).into())
@@ -48,34 +43,5 @@ impl Rule for ApplyToJoinRule {
 impl ApplyToJoinRule {
     pub fn create() -> BoxedRule {
         Box::new(ApplyToJoinRule {})
-    }
-}
-
-struct MaxOneRowVisitor();
-
-/// Return true if we can determine at most one row returns by the plan, otherwise false.
-impl PlanVisitor<bool> for MaxOneRowVisitor {
-    fn merge(a: bool, b: bool) -> bool {
-        a & b
-    }
-
-    fn visit_logical_agg(&mut self, plan: &LogicalAgg) -> bool {
-        plan.group_key().is_empty()
-    }
-
-    fn visit_logical_top_n(&mut self, plan: &LogicalTopN) -> bool {
-        plan.limit() == 1
-    }
-
-    fn visit_logical_expand(&mut self, plan: &LogicalExpand) -> bool {
-        plan.column_subsets().len() == 1 && self.visit(plan.input())
-    }
-
-    fn visit_logical_project_set(&mut self, plan: &LogicalProjectSet) -> bool {
-        plan.select_list().len() == 1 && self.visit(plan.input())
-    }
-
-    fn visit_logical_hop_window(&mut self, _plan: &LogicalHopWindow) -> bool {
-        false
     }
 }
